+++ src/icons/index.tsx
@@ -8,6 +8,8 @@ export {default as Rain} from "./rain"
 export {default as River} from "./river"
 export {default as Thunderstorm} from "./thunderstorm"
 export {default as Train} from "./train"
+export {default as Waves} from "./waves"
+export {default as Waterfall} from "./waterfall"
 
 export default function SVGWrapper(props: React.SVGProps<SVGElement>){
      return (